add & modularize dbQueries.ts
